Add user to team_members when create team.
